fix header

+++ uartParser.h
@@ -2,5 +2,5 @@ void copy_command(void);
 void process_command(void);
 void print_value (char *id, int value);
 void uart_ok(void);
-int parse_assignment (char input[16]);
+long parse_assignment (char input[16]);
 void process_uart(void);
